temp-24

+++ LIB.Module/modCameraVC0706_StateOperationImage.cpp
@@ -33,13 +33,13 @@ bool tCameraVC0706::tStateOperationImage::Go()
 		m_ImageReady = true;
 		m_pObj->OnImageReady(); // when picture is really exists
 
-		std::size_t ChunkSizeMax = m_Settings.ImageChunkSize / 4;
+		std::uint32_t ChunkSizeMax = m_Settings.ImageChunkSize / 4;
 		ChunkSizeMax *= 4;//it must be multiple of 4
 
 		const std::uint32_t ChunkDelay = m_Settings.ImageChunkDelayFromReq_us / 10;//in 0.01ms => 5000 / 10 = 500 => 50ms
 		const std::uint32_t ChunkDelay_ms = m_Settings.ImageChunkDelayFromReq_us / 1000;
 
-		std::size_t ChunkQty = FBufLen.Value / ChunkSizeMax;
+		std::uint32_t ChunkQty = FBufLen.Value / ChunkSizeMax;
 		if (FBufLen.Value % ChunkSizeMax)
 			++ChunkQty;
